Solution to Calculate Area and Perimeter of Rectangles

+++ Calculate Area and Perimeter of Rectangles@@67877d61dc107486ad5c274b/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 struct rectangle {
-    float l;
-    float b;
+    int l;
+    int b;
 };
 
 int main() {
@@ -15,7 +15,7 @@ int main() {
     }
 
     for (int i = 0; i < n; i++) {
-        printf("Rectangle %d: Area = %.2f, Perimeter = %.2f\n", i + 1, (r[i].l * r[i].b), 2 * (r[i].l + r[i].b));
+        printf("Rectangle %d: Area = %.2f, Perimeter = %.2f\n", i + 1, (float)(r[i].l * r[i].b), 2.0 * (r[i].l + r[i].b));
     }
 
     return 0;
